Initial state
3D ASCII in the PowerShell console.

Script written as part of my language learning. PowerShell was not intended for such use, but why not... the console does not have to be boring :)

The script shows the rotation of 5 different 3D objects such as a cube, triangle, pyramid, star and dodecahedron. Each of the objects is available by pressing the keys from '1' to '5'. The '0' key changes the line drawing character. The 'q' key terminates the script.